refactor: 마이페이지, 게시글 상세 페이지 구조 수정
